fix: Replace "page.expect_response" with "page.route"
The reason is that "page.expect_response" did lead to unhandled exceptions being
logged as errors on many tests. This is very problematic if a test fails,
because it would distract tracking down the real issue.

Using "page.route" it is also possible to intercept a response. The handler for
a route can explicitly decide when to fulfill the request.