feat: add upgrade tests
Adds two tests. One for the pre-upgrade where a user is created and
added to the `2FA Users` group. The user then logs in and sets up 2FA.
The TOTP secret as well as the username and password are written to a
file to be used by the post_upgrade test.

The post_upgrade test reads username, password and TOTP secret and then
logs in as the user with 2FA.

Issue: univention/dev/internal/team-nubus#1535